// src/types/station.ts
// Station raw data from CSV
export interface StationRawData {
  station_name: string;
  editor_x: string;
  editor_y: string;
  barcode_x: number;
  barcode_y: number;
  barcode_z: number;
  barcode_r: number;
  bay_name: string;
  station_type: string;
  port_id: string;
  port_type_code: string;
  direction_code: string;
  link_sc_id: string;
  buffer_size: string;
  mode_type: string;
  floor: string;
  zone_id: string;
  rail_index: string;
  sc_id: string;
  e84: string;
  teached: string;
  look_down: string;
  nearest_edge: string;
  nearest_edge_distance: string;
  eq_id: string;
}